fix regression in test
